Added PL version
